+++ repeat_test.go
@@ -1,62 +1,192 @@
 package repeat
 
 import (
+	"context"
 	"errors"
-	"fmt"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func testOperation(max int, result error) Operation {
-	c := 0
-	return func(e error) error {
-		defer func() { c++ }()
-		if c >= max && result == nil {
-			return &StopError{e}
-		}
-		if c >= max {
-			fmt.Println(result)
-			return result
-		}
+func TestCompose_NilInNilOut(t *testing.T) {
+	require.NoError(t, Compose(func(e error) error {
+		require.NoError(t, e)
 		return e
-	}
+	})(nil))
 }
 
-func TestRepeat(t *testing.T) {
-	op10 := func(e error, c int) error {
-		defer func() { c++ }()
-		if c == 0 {
-			assert.NoError(t, e)
-		} else {
-			assert.Error(t, e, "temporary")
-		}
-		if c >= 10 {
-			return nil
-		}
-		return &TemporaryError{errors.New("temporary")}
-	}
-	assert.NoError(t, Repeat(FnWithErrorAndCounter(op10), testOperation(10, nil)))
-	assert.EqualError(t, Repeat(FnWithErrorAndCounter(op10), testOperation(5, nil)), "temporary")
-	assert.EqualError(t, Repeat(FnWithErrorAndCounter(op10), testOperation(5, errors.New("external"))), "external")
+func TestCompose_ErrInErrOut(t *testing.T) {
+	require.EqualError(t, Compose(func(e error) error {
+		require.EqualError(t, e, "oil")
+		return e
+	})(errors.New("oil")), "oil")
 }
 
-func TestCompose(t *testing.T) {
-	op10 := func(e error, c int) error {
-		defer func() { c++ }()
-		if c == 0 {
-			assert.NoError(t, e)
-		} else {
-			assert.Error(t, e, "temporary")
-		}
-		if c >= 10 {
+func TestCompose_TemporaryErrInTemporaryErrOut(t *testing.T) {
+	require.EqualError(t, Compose(func(e error) error {
+		require.EqualError(t, e, "repeat.temporary: cat")
+		return HintTemporary(e)
+	})(HintTemporary(errors.New("cat"))), "repeat.temporary: cat")
+}
+
+func TestCompose_ErrInStopErrOut(t *testing.T) {
+	require.EqualError(t, Compose(func(e error) error {
+		return HintStop(e)
+	})(errors.New("bob")), "repeat.stop: bob")
+}
+
+func TestCompose_NillOverridesTemporaryError(t *testing.T) {
+	require.NoError(t, Compose(
+		func(e error) error {
+			return HintTemporary(e)
+		},
+		Done,
+	)(errors.New("ann")))
+}
+
+func TestCompose_OverrideTemporaryError(t *testing.T) {
+	require.EqualError(t, Compose(
+		func(e error) error {
+			return HintTemporary(e)
+		},
+		func(e error) error {
+			return HintTemporary(errors.New("pong"))
+		},
+	)(errors.New("ping")), "repeat.temporary: pong")
+}
+
+func TestOnce_NonTemporaryErrOut(t *testing.T) {
+	require.EqualError(t, Once(func(e error) error {
+		return HintTemporary(errors.New("zed"))
+	}), "zed")
+}
+
+func TestOnce_NonStopErrOut(t *testing.T) {
+	require.NoError(t, Once(func(e error) error {
+		return HintStop(nil)
+	}))
+}
+
+func TestOnce_ErrOut(t *testing.T) {
+	require.EqualError(t, Once(func(e error) error {
+		return errors.New("aim")
+	}), "aim")
+}
+
+func TestFnRepeat_NilInStopErrWithNilOut(t *testing.T) {
+	require.EqualError(t, FnRepeat(
+		func(e error) error {
+			require.NoError(t, e)
+			return e
+		},
+		StopOnSuccess(),
+	)(nil), "repeat.stop")
+}
+
+func TestFnRepeat_ErrInErrOut(t *testing.T) {
+	require.EqualError(t, FnRepeat(
+		func(e error) error {
+			require.EqualError(t, e, "oil")
+			return e
+		},
+	)(errors.New("oil")), "oil")
+}
+
+func TestFnRepeat_TemporaryErrInSameStopErrOut(t *testing.T) {
+	require.EqualError(t, FnRepeat(
+		LimitMaxTries(1),
+		func(e error) error {
+			require.EqualError(t, e, "repeat.temporary: cat")
+			return HintTemporary(e)
+		},
+	)(HintTemporary(errors.New("cat"))), "repeat.stop: cat")
+}
+
+func TestFnRepeat_ErrInStopErrOut(t *testing.T) {
+	require.EqualError(t, FnRepeat(func(e error) error {
+		return HintStop(e)
+	})(errors.New("bob")), "repeat.stop: bob")
+}
+
+func TestRepeat_WithNoErrors(t *testing.T) {
+	cn := 0
+	require.NoError(t, Repeat(
+		LimitMaxTries(3),
+		FnWithErrorAndCounter(func(e error, c int) error {
+			defer func() { cn++ }()
+			require.Equal(t, cn, c, "should be equal on every")
+
+			switch {
+			case c < 3:
+				require.NoError(t, e, "no error on every call")
+			default:
+				require.Fail(t, "cant be here, only three tries")
+			}
+
 			return nil
-		}
-		return &TemporaryError{errors.New("temporary")}
-	}
-	assert.NoError(t, Repeat(Compose(FnWithErrorAndCounter(op10), testOperation(10, nil))))
-	assert.EqualError(t, Repeat(Compose(FnWithErrorAndCounter(op10), testOperation(5, nil))), "temporary")
-	assert.EqualError(t, Repeat(Compose(FnWithErrorAndCounter(op10), testOperation(5, errors.New("external")))), "external")
+		}),
+	))
+	require.Equal(t, 3, cn)
+}
+
+func TestRepeat_WithTemporaryErrors(t *testing.T) {
+	cn := 0
+	require.EqualError(t, Repeat(
+		LimitMaxTries(3),
+		// Should be called three times until LimitMaxTries stops the execution.
+		FnWithErrorAndCounter(func(e error, c int) error {
+			defer func() { cn++ }()
+			require.Equal(t, cn, c, "should be equal on every")
+
+			switch c {
+			case 0:
+				require.NoError(t, e,
+					"no error on first call (started with no error)")
+			case 1, 2:
+				require.EqualError(t, e, "repeat.temporary: my temporary",
+					"the same error func returns")
+			default:
+				require.Fail(t, "cant be here")
+			}
+
+			return HintTemporary(errors.New("my temporary"))
+		}),
+	), "my temporary")
+	require.Equal(t, 3, cn)
+}
+
+func TestRepeat_WithErrors(t *testing.T) {
+	cn := 0
+	require.EqualError(t, Repeat(
+		FnWithErrorAndCounter(func(e error, c int) error {
+			defer func() { cn++ }()
+			require.Equal(t, cn, c, "should be equal on every")
+
+			if c == 2 {
+				return errors.New("my real")
+			}
+
+			return HintTemporary(errors.New("my temporary"))
+		}),
+	), "my real")
+	require.Equal(t, 3, cn)
+}
+
+func TestRepeat_WithStopErrors(t *testing.T) {
+	cn := 0
+	require.EqualError(t, Repeat(
+		FnWithErrorAndCounter(func(e error, c int) error {
+			defer func() { cn++ }()
+			require.Equal(t, cn, c, "should be equal on every")
+
+			if c == 2 {
+				return HintStop(errors.New("my real"))
+			}
+
+			return HintTemporary(errors.New("my temporary"))
+		}),
+	), "my real")
+	require.Equal(t, 3, cn)
 }
 
 func TestWrap(t *testing.T) {
@@ -69,8 +199,8 @@ func TestWrap(t *testing.T) {
 		}
 	}
 
-	assert.NoError(t, Wrap(wr).Compose(Nope, Nope)(nil))
-	assert.Equal(t, 2, c)
+	require.NoError(t, Wrap(wr).Compose(Nope, Nope)(nil))
+	require.Equal(t, 2, c, "wr called two times according to number of ops in Compose")
 }
 
 func TestCpp_C_D(t *testing.T) {
@@ -81,8 +211,8 @@ func TestCpp_C_D(t *testing.T) {
 		return e
 	}
 
-	assert.NoError(t, Cpp(cd, cd).Compose(Nope)(nil))
-	assert.Equal(t, 2, c)
+	require.NoError(t, Cpp(cd, cd).Compose(Nope)(nil))
+	require.Equal(t, 2, c)
 }
 
 func TestCpp_C_NoD(t *testing.T) {
@@ -93,8 +223,8 @@ func TestCpp_C_NoD(t *testing.T) {
 		return e
 	}
 
-	assert.EqualError(t, Cpp(cd, cd).Compose(Nope)(errGolden), errGolden.Error())
-	assert.Equal(t, 1, c)
+	require.EqualError(t, Cpp(cd, cd).Compose(Nope)(errGolden), errGolden.Error())
+	require.Equal(t, 1, c)
 }
 
 func TestCpp_C_ErrOP_D(t *testing.T) {
@@ -109,8 +239,8 @@ func TestCpp_C_ErrOP_D(t *testing.T) {
 		return errGolden
 	}
 
-	assert.EqualError(t, Cpp(cd, cd).Compose(errOp)(nil), errGolden.Error())
-	assert.Equal(t, 2, c)
+	require.EqualError(t, Cpp(cd, FnDone(cd)).Compose(errOp)(nil), errGolden.Error())
+	require.Equal(t, 2, c)
 }
 
 func TestCpp_C_PanicOP_D(t *testing.T) {
@@ -125,8 +255,61 @@ func TestCpp_C_PanicOP_D(t *testing.T) {
 		panic(errGolden)
 	}
 
-	assert.Panics(t, func() {
+	require.Panics(t, func() {
 		Cpp(cd, cd).Compose(errOp)(nil)
 	})
-	assert.Equal(t, 2, c)
+	require.Equal(t, 2, c)
+}
+
+func TestCpp_C_Op_ErrorD(t *testing.T) {
+	c := 0
+
+	cc := func(e error) error {
+		c++
+		return e
+	}
+
+	dd := func(error) error {
+		c++
+		return errGolden
+	}
+
+	require.Panics(t, func() {
+		Cpp(cc, dd).Compose(Nope)(nil)
+	})
+	require.Equal(t, 2, c)
+}
+
+func TestCpp_TransparentC(t *testing.T) {
+	c := 0
+
+	cc := func(e error) error {
+		c++
+		require.EqualError(t, Cause(e), errGolden.Error())
+		return nil
+	}
+
+	dd := func(e error) error {
+		c++
+		require.NoError(t, Cause(e))
+		return nil
+	}
+
+	op := func(e error) error {
+		c++
+		require.EqualError(t, Cause(e), errGolden.Error())
+		return nil
+	}
+
+	require.NoError(t, Cause(
+		Cpp(cc, dd).Compose(op)(HintTemporary(errGolden)),
+	))
+	require.Equal(t, 3, c)
+}
+
+func TestRepeatWithContext(t *testing.T) {
+	ctx, cancel := context.WithCancel(context.Background())
+	cancel()
+
+	require.EqualError(t, WithContext(ctx).Once(Nope), "context canceled")
 }