+++ CHANGELOG.md
@@ -24,3 +24,10 @@
 # 1.3
 
 * add: Once and FnRepeat to Repeater and global functions
+
+# 1.4
+
+* fix: Once calls global compose
+* ref: some changes in Cpp concept. It is transparent for input errors now, it also panics if D fails
+* add: Done, FnDone, FnOnlyOnce
+* add: 100% test coverage